requirements.txt
pytest
selenium
webdriver-manager
requests
allure-pytest